Fix "detected dubious ownership in repository"
Files in docker container are all affected to the non root user (your
host user).
But when running `composer install` or `composer update`, the script,
running as root within docker, detects that the owner of the files is
not the runner of the script and display a warning as follow:

> The repository at "/app" does not have the correct ownership and git
> refuses to use it:
>
> fatal: detected dubious ownership in repository at '/app'
> To add an exception for this directory, call:
>       git config --global --add safe.directory /app

This commit fixes that by trusting the /app folder for git **inside**
the docker container.