DX: Add debug make target and documentation
Debugging is always a pain to setup and it's what makes most developer
not use it.
This provides a working way to say it up on PHPStorm as well as a make
target to ease switching between a container with XDEBUG on and a
container with XDEBUG off.

Most of the credits are to the developers on dunglas/symfony-docker.