Remove APP_SECRET env var
This value is never used and triggers a warning from GitGuardian (tool
from Github detecting publicly exposed secrets).
It wasn't a risky secret as it would only affect dev machines but after
digging, it won't be used in this project and can be safely removed.

See https://symfony.com/blog/new-in-symfony-7-2-optional-secret for more
information