Use an image prefix to prevent conflict
By default, the project build a docker image named app-php but this can
conflict with other images built in other project.
So now, we use an image prefix to make sure it's different per project.
If the image should be shared between projects, we can edit this prefix
locally in the .env.local file